backend/requirements.txt
Django==5.2
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
flower==2.0.1
gunicorn==22.0.0
whitenoise==6.6.0

# REST API
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-allauth==0.63.3

# GraphQL
graphene-django==3.2.0

# Security
django-extensions==3.2.3
django-axes==6.4.0
django-ratelimit==4.1.0
PyJWT==2.8.0
cryptography==42.0.5

# Monitoring
django-prometheus==2.3.1

# Environment
django-environ==0.11.2

# WebSocket
channels==4.1.0
channels-redis==4.2.0

# Email
django-ses==3.5.2

# Development
django-debug-toolbar==4.4.2

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0